Separate out a Daydream backend
+ some viewManager refactor

GearVRf-DCO-1.0-Signed-off-by: Mihail Marinov <m.marinov@samsung.com>
