+++ GVRf/Framework/backend_oculus/src/main/java/org/gearvrf/OvrActivityHandler.java
@@ -10,4 +10,6 @@ interface OvrActivityHandler {
     public boolean onBack();
 
     public boolean onBackLongPress();
+
+    void setViewManager(GVRViewManager viewManager);
 }
