+++ GVRf/Framework/backend_oculus/src/main/jni/ovr_view_manager.cpp
@@ -13,10 +13,9 @@
  * limitations under the License.
  */
 
-#include "ovr_activity.h"
-#include "ovr_view_manager.h"
 #include "../engine/renderer/renderer.h"
 #include "../objects/components/camera.h"
+#include "objects/textures/render_texture.h"
 
 #include <jni.h>
 
@@ -24,55 +23,14 @@ namespace gvr {
 
 extern "C" {
 
-void Java_org_gearvrf_OvrViewManager_cull(JNIEnv * jni, jclass clazz,
-        jlong jscene, jlong jcamera, jlong jshader_manager) {
-    Scene* scene = reinterpret_cast<Scene*>(jscene);
-    Camera* camera = reinterpret_cast<Camera*>(jcamera);
-    ShaderManager* shader_manager = reinterpret_cast<ShaderManager*>(jshader_manager);
-    gRenderer = Renderer::getInstance();
-    gRenderer->cull(scene, camera, shader_manager);
-
-}
-
-
-void Java_org_gearvrf_OvrViewManager_makeShadowMaps(JNIEnv * jni, jclass clazz,
-        jlong jscene, jlong jshader_manager, jint width, jint height) {
-    Scene* scene = reinterpret_cast<Scene*>(jscene);
-
-    ShaderManager* shader_manager = reinterpret_cast<ShaderManager*>(jshader_manager);
-    gRenderer = Renderer::getInstance();
-    gRenderer->makeShadowMaps(scene, shader_manager, width, height);
-}
-
-void Java_org_gearvrf_OvrViewManager_renderCamera(JNIEnv * jni, jclass clazz,
-        jlong appPtr, jlong jscene, jlong jcamera, jlong jshader_manager,
-        jlong jpost_effect_shader_manager, jlong jpost_effect_render_texture_a,
-        jlong jpost_effect_render_texture_b) {
-    GVRActivity *activity = (GVRActivity*)appPtr;
-
-    Scene* scene = reinterpret_cast<Scene*>(jscene);
-    Camera* camera = reinterpret_cast<Camera*>(jcamera);
-    ShaderManager* shader_manager =
-            reinterpret_cast<ShaderManager*>(jshader_manager);
-    PostEffectShaderManager* post_effect_shader_manager =
-            reinterpret_cast<PostEffectShaderManager*>(jpost_effect_shader_manager);
-    RenderTexture* post_effect_render_texture_a =
-            reinterpret_cast<RenderTexture*>(jpost_effect_render_texture_a);
-    RenderTexture* post_effect_render_texture_b =
-            reinterpret_cast<RenderTexture*>(jpost_effect_render_texture_b);
-
-    activity->viewManager_.renderCamera(scene, camera,
-            shader_manager, post_effect_shader_manager,
-            post_effect_render_texture_a, post_effect_render_texture_b);
-}
-
-void Java_org_gearvrf_OvrViewManager_readRenderResultNative(JNIEnv * jni,
-        jclass clazz, jlong jrender_texture, jobject jreadback_buffer) {
+void Java_org_gearvrf_OvrViewManager_readRenderResultNative(JNIEnv *jni,
+                                                            jclass clazz, jlong jrender_texture,
+                                                            jobject jreadback_buffer) {
 
     uint8_t *pReadbackBuffer = (uint8_t *) jni->GetDirectBufferAddress(
             jreadback_buffer);
-    RenderTexture* render_texture =
-            reinterpret_cast<RenderTexture*>(jrender_texture);
+    RenderTexture *render_texture =
+            reinterpret_cast<RenderTexture *>(jrender_texture);
     int width = render_texture->width();
     int height = render_texture->height();
 
@@ -85,11 +43,11 @@ void Java_org_gearvrf_OvrViewManager_readRenderResultNative(JNIEnv * jni,
     GLuint renderTextureFBO = render_texture->getFrameBufferId();
     glBindFramebuffer(GL_DRAW_FRAMEBUFFER, renderTextureFBO);
     glBlitFramebuffer(0, 0, width, height, 0, 0, width, height,
-            GL_COLOR_BUFFER_BIT, GL_NEAREST);
+                      GL_COLOR_BUFFER_BIT, GL_NEAREST);
     glBindFramebuffer(GL_READ_FRAMEBUFFER, renderTextureFBO);
 
     glReadPixels(0, 0, width, height, GL_RGBA, GL_UNSIGNED_BYTE,
-            pReadbackBuffer);
+                 pReadbackBuffer);
 
 // recover FBO bindings
     glBindFramebuffer(GL_READ_FRAMEBUFFER, currentReadFBO);
@@ -98,52 +56,4 @@ void Java_org_gearvrf_OvrViewManager_readRenderResultNative(JNIEnv * jni,
 
 } // extern "C"
 
-//=============================================================================
-//                             GVRViewManager
-//=============================================================================
-
-GVRViewManager::GVRViewManager() {
-}
-
-GVRViewManager::~GVRViewManager() {
-}
-
-void GVRViewManager::renderCamera(Scene* scene,
-        Camera* camera, ShaderManager* shader_manager,
-        PostEffectShaderManager* post_effect_shader_manager,
-        RenderTexture* post_effect_render_texture_a,
-        RenderTexture* post_effect_render_texture_b) {
-#ifdef GVRF_FBO_FPS
-    // starting to collect rendering time
-    // first flash GPU tasks
-    glFinish();
-
-    // collect data
-    struct timeval start, end;
-    double t1, t2;
-    float fps = 0.0;
-    gettimeofday(&start, NULL);
-#endif
-    gRenderer->renderCamera(scene, camera, shader_manager,
-                post_effect_shader_manager, post_effect_render_texture_a,
-                post_effect_render_texture_b);
-
-#ifdef GVRF_FBO_FPS
-    // finish rendering
-    glFinish();// force rendering
-
-    gettimeofday(&end, NULL);
-    t1 = start.tv_sec + (start.tv_usec / 1000000.0);
-    t2 = end.tv_sec + (end.tv_usec / 1000000.0);
-    gTotalSec += (t2 - t1);
-    gNumFrame++;
-
-    fps = (float)gNumFrame/gTotalSec;
-    if(!(gNumFrame%50) )
-    {
-        LOGI("FPS is %.2f gNumFrame=%d, gTotalSec = %.2f", fps, gNumFrame, gTotalSec);
-    }
-#endif
-
-}
 }
