+++ GVRf/Framework/settings.gradle
@@ -1,4 +1,5 @@
 include ':framework'
 if (!hasProperty("ARM64")) {
     include ':backend_oculus'
+    include ':backend_daydream'
 }
